Move brief API documentation out of example
Really love this component and thought it might be helpful for other people if the props for a `<Draggable/>` were more clearly documented. I don't usually look in code examples for documentation about a given API and it might help people clearly see how to clearly use the module while still including an easy-to-get usage. Open to thoughts / feedback / &c.!