Use explicit constructor in Draggable to fix IE10 null reference
There's a problem in IE10 with ES2015 classes and the super class.
Essentially the prototype isn't set up correctly which leads to
this.props being null when setting up position in the construction.

However, props are passed to the class constructor explicitly by React,
so we can use them from there.

Tested by running the examples in IE 10 on Win 7.